feat: iOS xcframework build target
Adds ios-build-target branch parity with Solar-Punk-Ltd/bee-lite-java's
fork. Produces build/Mobile.xcframework via `make build-ios`, containing
both bee-lite (Swarm) and kubo (IPFS) in a single Go runtime — required
because two gomobile-bound xcframeworks cannot coexist in one iOS
process (they fight over Go runtime TLS slots and crash at startup).

Changes mirror bee-lite-java@ios-build-target's iOS-readiness fixes:

- mobile/types.go: StampData.{BatchDepth,BucketDepth} byte→int32.
  gomobile emits `byte` for uint8, which isn't a valid Obj-C type
  (known bug in golang.org/x/mobile/bind/genobjc.go). int32 sidesteps it.
- mobile/upload.go: rLevel byte→int32 (UploadManager.Upload,
  getRedundancyLevel).
- mobile/mobile-wrapper.go: Upload interface and impl unified on
  `rLevel int32`. Obj-C selectors encode parameter names — iface↔impl
  mismatch produces distinct selectors and breaks protocol conformance.
- mobile/stamp-manager.go: int32() casts on batch.Depth() and
  batch.BucketDepth().
- Makefile: +build-ios, +verify-ios.
- go.mod: replace ethersphere/bee/v2 => flotob/bee/v2 v2.7.0-freedom.2.
  Pulls patched bee with kademlia metric storage close and wrapped-index
  cache close fixes that haven't been upstreamed yet. Keep in lockstep
  with bee-lite-java's same replace.

Verified: make probe-co green (bee+kubo one process, kubo 39 mainnet
peers); make build-ios green (533MB xcframework, 179MB device slice);
flotob/bee/v2 string present in the compiled binary.

+++ mobile/stamp-manager.go
@@ -30,8 +30,8 @@ func (sm *StampManager) GetAllBatches() {
 			Label:         batch.Label(),
 			BatchIdHex:    hex.EncodeToString(batch.ID()),
 			BatchAmount:   batch.Amount().String(),
-			BatchDepth:    batch.Depth(),
-			BucketDepth:   batch.BucketDepth(),
+			BatchDepth:    int32(batch.Depth()),
+			BucketDepth:   int32(batch.BucketDepth()),
 			ImmutableFlag: batch.ImmutableFlag(),
 		}
 	}
